Address MSVS ptr overflow warnings when using negative image origin

+++ src/openexr.imageio/exrinput.cpp
@@ -1351,8 +1351,8 @@ OpenEXRInput::read_native_tiles(int subimage, int miplevel, int xbegin,
         tmpbuf.resize(nxtiles * nytiles * m_spec.tile_bytes(true));
         data = tmpbuf.data();
     }
-    char* buf = (char*)data - xbegin * pixelbytes
-                - ybegin * pixelbytes * m_spec.tile_width * nxtiles;
+    char* buf = (char*)data - ptrdiff_t(xbegin * pixelbytes)
+                - ptrdiff_t(ybegin * pixelbytes * m_spec.tile_width * nxtiles);
 
     try {
         Imf::FrameBuffer frameBuffer;
